refactor(pack): refactor hosted email
