+++ packages/manager/apps/telecom/src/app/telecom/pack/slots/hostedEmail/add/add.routing.js
@@ -1,0 +1,11 @@
+import template from './add.html';
+import controller from './add.controller';
+
+export default /* @ngInject */ ($stateProvider) => {
+  $stateProvider.state('telecom.packs.pack.hostedEmail-add', {
+    url: '/hostedEmail/add',
+    template,
+    controller,
+    controllerAs: 'ctrl',
+  });
+};
