--- packages/manager/apps/telecom/src/app/telecom/pack/slots/hostedEmail/add/pack-hostedEmail-add.js
@@ -1,10 +1,0 @@
-angular.module('managerApp').config(($stateProvider) => {
-  $stateProvider.state('telecom.packs.pack.hostedEmail-add', {
-    url: '/hostedEmail/add',
-    templateUrl:
-      'app/telecom/pack/slots/hostedEmail/add/pack-hostedEmail-add.html',
-    controller: 'PackHostedEmailAddCtrl',
-    controllerAs: 'ctrl',
-    translations: { value: ['.'], format: 'json' },
-  });
-});
