+++ packages/manager/apps/telecom/src/app/telecom/pack/slots/hostedEmail/hostedEmail.html
